% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIDs.R
\name{assignIDs}
\alias{assignIDs}
\title{Assign ID numbers to samples}
\usage{
assignIDs(data, xCoord = "x", yCoord = "y",
  neighborhood = "tertiary")
}
\arguments{
\item{data}{The dataset containing x- and y-coordinates.}

\item{xCoord}{A character string specifying the name of the x-coordinate
variable.}

\item{yCoord}{A character string specifying the name of the y-coordinate
variable.}

\item{neighborhood}{A character string specifying the neighborhood
structure to be considered.  The options are "primary", "secondary", and
"tertiary".}
}
\value{
The dataset appended with a variable called "assignID".  This
variable provides the assigned IDs as a numeric variable.
}
\description{
\code{assignIDs} attempts to assign ID numbers to contiguous regions.
}
\examples{
data("TAMdata")
TAMdata <- assignIDs(TAMdata)
}
