% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseStructures_final.R
\name{chooseStructures}
\alias{chooseStructures}
\title{Choose support structure(s)}
\usage{
chooseStructures(rangeObj, cDist = 0.1, sdWithin = 1,
  defaultStructure = "nextHighest", sdDefault = 1,
  thresholdNumSup = 0.5, sdElim = 2, noZeroRangeStructs = TRUE)
}
\arguments{
\item{rangeObj}{An object of class \code{rangeList}.  This object is a list
that contains the estimated range parameter(s).  This is obtained by a call
to the \code{\link{estRange}} function.}

\item{cDist}{A distance expressing how far beyond the limits of the data
that the support sites should be placed.  In the \code{rScale} function,
the data are rescaled such that data locations are a distance of 1 apart.
Therefore, a cDist of 0.5 is half the distance between data points.}

\item{sdWithin}{The maximum allowed distance between any data point and
its nearest support site (in terms of the number of standard deviations
of the smoothing kernel).  The default is 1, meaning that data must be
within one standard deviation of the smoothing kernel of a support site.
Smaller values of sdWithin lead to demser support structures.  This value
should never be set higher than 1.}

\item{defaultStructure}{The default structure if the number of support
points exceeds the proportion threshold of data set by
\code{thresholdNumSup}.  The options are '5x' or 'nextHighest'.}

\item{sdDefault}{The default standard deviation of the smoothing kernel.
The default standard deviation is used if (1) the number of support sites
exceeds that allowed by \code{thresholdNumSup} or (2) the recommended
standard deviation of the smoothing kernel exceeds the largest distance
between data points.  \code{sdDefault} is a value mutliplied by the maximum
observed distance between any data point and the nearest support site.  For
an \code{sdDefault} of 1, the maximum distance is mutliplied by 1, and that
value is used for the standard deviation of the smoothing kernel when a
default support structure is used.}

\item{thresholdNumSup}{A threshold that limits the number of support sites
allowed for computational efficiency.  The threshold is a proportion of the
observations.  A threshold of 0.5 limits the number of allowed support sites
to 50\% of the number of observations.}

\item{sdElim}{A threshold for eliminating support sites (in terms of the
standard deviation of the smoothing kernel).  Support sites are removed if
they are not within the specified number of standard deviations of a data
point.  This is used to remove support sites for irregularly shaped samples.}

\item{noZeroRangeStructs}{A TRUE/FALSE argument specifying how to handle
instances in which the estimated range parameter is 0.  If
\code{noZeroRangeStructs=TRUE} and the estimated range parameter is 0, then
no support structure will be used for the corresponding data.  If a spatial
variable was provided, then no support structure will be used for the data
in the corresponding level of that variable.  If no spatial variable was
provided, then a support structure will not be used at all.  Instead, a
raster-level intercept will be incorporated to account for extra noise.
If \code{noZeroRanges=FALSE} and the estimated range parameter is 0, then a
default support structure will be used.}
}
\value{
An object of class \code{structureList} that includes the
information below.
\describe{
  \item{\code{structure}}{A list of information about the support
  structures.  If a spatial variable is provided, then there will be a list
  for each level of that variable.  The structure list includes
  \code{recommendStruct} (the recommended support structure),
  \code{recommendSd} (the recommended standard deviation of the smoothing
  kernel), \code{nSupportReduced} (the number of support sites after
  removing those more than \code{sdElim} standard deviations away from a
  data point), \code{coordsU} (a data frame of the coordinates of the
  support sites, by sample), \code{coordsData} (a data frame of the
  coordinates of the data), \code{defaultStatus} (a character string
  informing the user if a default support structure was chosen), and
  \code{nRowSupport} (the number of rows of support sites for alternating
  support structures).}
  \item{\code{data}}{The dataset, appended with new x- and y-coordinates.}
  \item{\code{subjectVar}}{A character string denoting the subject
  variable.}
  \item{\code{sampleVar}}{A character string denoting the sample variable.}
  \item{\code{spatialVar}}{A character string denoting the spatial
  variable.}
  \item{\code{outcome}}{A character string denoting the outcome of
  interest.}
}
}
\description{
The \code{chooseStructures} function selects support structures for every
level of the spatial variable (\code{spatialVar} argument) specified in the
\code{\link{estRange}} function.  If no spatial variable is provided, then
only a single set of support structures is chosen, one structure for each
sample.  See the \strong{Details} section for more information on how the
structures are selected.
}
\details{
The process convolution approach uses points called support
sites that help to account for underlying spatial structure in data.
Collections of support sites are called support structures.
How many and where to place support sites were questions that drove the
research leading to the creation of this function.

The \code{chooseStructures} function uses the procedure detailed below.
If a spatial variable was provided in the \code{\link{estRange}} function,
then the procedure is performed for every level of the spaital variable.

The procedure starts by building all of the fixed support structures.  The
fixed support structures are a group of five support structures ranging
from five to twelve support sites.  The minimum distance between each data
point and the closest support site is measured for all observations.
Support sites that are more than (\code{sdElim})*(recommended standard
deviation of smoothing kernel) are removed.  The total number of support
sites is counted across all samples, and each structure is checked to see
if all data are within (\code{sdWithin})*(recommended standard deviation
of smoothing kernel).

Once the fixed support structures are created, an iterative loop is used to
generate the alternating support structures.  These are support structures
in which the rows of support sites have alternating numbers of support
sites.  For each support, the support structure is built and support sites
more than (\code{sdElim})*(recommended standard deviation of smoothing
kernel) are removed.  The support structure is then checked to see if (1)
all the data are within (\code{sdWithin})*(recommended standard deviation
of smoothing kernel) and (2) the number of support sites across all samples
has exceeded (\code{thresholdNumSup})*(total number of observations).  If
either criteria is met, the iteratie loop stops.  The alternating
structures are compared to the fixed structures, and the support structure
with the fewest support sites in which all of the data are within
(sdWithin)*(recommended standard deviation of smoothing kernel) is chosen
as the support structure.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
structs <- chooseStructures(rangs)
}
