% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAUCdata_final.R
\name{createAUCData}
\alias{createAUCData}
\title{Calculate the area under the curve (AUC) across m/z values on the raster
level.}
\usage{
createAUCData(data, mzThreshold = 0.3, usedMergeIMSFiles = TRUE)
}
\arguments{
\item{data}{A dataset containing columns of m/z measurements.}

\item{mzThreshold}{An m/z threshold that determines which m/z values
are combined when calculating the AUC.  See \code{Details} for more
information on how this works.}
}
\value{
A new dataset with the AUC calculations.  Any columns with column
names that do not start with an "X" and then a nubmer will be kept.  All
original m/z columns will be removed.
}
\description{
\code{createAUCData} calculates the AUC for sets of m/z values.
}
\details{
This function is designed to work with data exported from
SCiLS Lab software.  When using SCiLS make sure to export the m/z values
as columns, meaning every column should represent a separate m/z value,
and every column name is the m/z value When the data are imported into R,
R will add an 'X' in front of each number.  Do not remove the 'X' in front
as this function will automatically do that.

The \code{createAUCData} first searches through the m/z values and
determines which to combine based on the \code{mzThreshold}.  The function
goes one-by-one through the m/z values and determines which other
fragments are within the \code{mzThreshold}  Those within the mzThreshold
are considered to represent the same fragment and are used to
calculate the AUC.

The \code{mzThreshold} must be smaller than the smallest m/z distance
between m/z values representing different fragments.  For example,
if the highest m/z representing fragment 1 is 2466.005859, and the
smallest m/z representing fragment 2 is 2466.878418, then the threshold
must be less than 0.872559 (2466.878418-2466.005859=0.872559) to
correctly distinguish between the two fragments.

As a result of the way m/z values are combined, this function
should be used for data in which the m/z distances defining each
fragment are less than the m/z distances between fragments.  This is often
the case when the data that have gone through the peak picking
process.  If entire spectra for spot-level data are submitted to this
function, all m/z values will be combined.
}
