% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRange_final.R
\name{estRange}
\alias{estRange}
\title{Estimate the range parameter}
\usage{
estRange(rScaleObj, outcome, spatialVar = NULL, semivEst = "modulus",
  logTransform = TRUE, covarianceModel = "gaussian")
}
\arguments{
\item{rScaleObj}{An object of class \code{rScaleList}.  This object is the
result of using the \code{\link{rScale}} function.}

\item{outcome}{A character string specifying the outcome of interest.  This
is the variable that will later be modeled.}

\item{spatialVar}{An optional character string specifying a binary or
categorical variable.  If a variable is input, the range will be estimated
for all levels of that variable.}

\item{semivEst}{The form of the semivariance estimator.  The options are
'classical' and 'modulus'.  The classical estimator is the method of
moments etimator, and the modulus estimator is robust estimator from
Hawkins and Cressie.}

\item{logTransform}{A TRUE/FALSE variable indicating whether or not the
outcome should be log-transformed.  Imaging mass spectrometry data are
typically lognormally distributed, and so the default is TRUE.}

\item{covarianceModel}{A character string specifying the form of the
covariance model.  The only current option is 'gaussian'.}
}
\value{
A list of class \code{rangeList} containing the estimated range
and other information supplied to the \code{estRange} function.

\describe{
  \item{\code{data}}{A data frame containing the data.}
  \item{\code{subjectVar}}{A character string denoting the subject
  variable.}
  \item{\code{sampleVar}}{A character string denoting the sample variable.}
  \item{\code{spatialVar}}{A character string denoting the spatial
  variable.}
  \item{\code{outcome}}{A character string denoting the outcome of
  interest.}
  \item{\code{estRange}}{A single value or vector of values representing
  the estimated range parameter.  If no spatial variable is given then
  this will be a single value.  If a spatial variable is given then this
  will be a vector of values, one for each level of the spatial variable.
  Each estimated range will be named for its corresponding spatial
  variable level.}
  \item{\code{estSig2}}{A single value or vector of values representing
  the estimated variance parameter \eqn{\sigma^2}.  The parameter
  \eqn{\sigma^2} is used to calculate the covariance function.  For more
  information, see the \code{\link[geoR]{cov.spatial}} function in the
  \code{geoR} package.  If no spatial variable is given then this will be
  a single value.  If a spatial variable is given then this will be a
  vector of values, one for each level of the spatial variable.  Each
  estimated range will be named for its corresponding spatial variable
  level.}
  \item{\code{semivarFit}}{The empirical variogram.  This is a result of
  a call to the \code{\link[geoR]{variog}} function in the \code{geoR}
  package.  If no spatial variable is provided, then this is a single
  object.  If a spatial variable is provided, then this is a list of
  objects, one object for every level of the spatial variable.}
  \item{\code{covModelFit}}{The fitted covariance model.  This is a result
  of a call to the \code{\link[geoR]{variofit}} function in the \code{geoR}
  package.  If no spatial variable is provided, then this is a single
  object.  If a spatial variable is provided, then this is a list of
  objects, one object for every level of the spatial variable.}
}
}
\description{
The \code{estRange} function is a wrapper for the
\code{\link[geoR]{variog}} and \code{\link[geoR]{variofit}} functions in
the \code{geoR} package.  \code{estRange} calculates the semivariance at
every observed distance within samples in the rescaled data and then fits
those estimates to a covariance model.
}
\examples{
data("TAMdata")
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
}
\references{
Ribeiro, Jr., PJ and Diggle, PJ. 2018. geoR: Analysis of
Geostatistical Data. R package version 1.7-2.1.

Cressie, N and Hawkins, DM. 1980. Robust estimation of the
variogram: I. \emph{Journal of the International Association for
Mathematical Geology}, 12(2):115-125.
}
