% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSCiLS_final.R
\name{mergeIMSFiles}
\alias{mergeIMSFiles}
\title{Merge spot-level data and region spots from SCiLS}
\usage{
mergeIMSFiles(spotData, regionSpots, trimTops = TRUE)
}
\arguments{
\item{spotData}{A dataset containing the spot-level data in which the
columns represent m/z values, and the rows represent spots.}

\item{regionSpots}{A dataset containing the region spots.  This should be
an exported file that only contains the spot identifiers and
the x- and y-coordinates.}

\item{trimTops}{A TRUE/FALSE variable indicating whether or not the top
rows of each dataset should be trimmed to remove metadata before merging.
If metadata exists in the top rows then trimTops should be set to TRUE
to remove it before proceeding to the next step.  Otherwise the metadata
will ineterfere with downstream steps.}
}
\value{
A merged data frame with spot-level data and corresponding
coordinates.
}
\description{
\code{mergeIMSFiles} merges the two files needed to create spot-level data with
coordinates.  The first is the spot-level data without coordinates, and the
second is the region spots with coordinates.  We will use the terms spots and
rasters interchangeably.  Though data are collected over circular areas
dependent on the laser diameter, the rasterization process makes data
representation easier.
}
