% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic functions.R
\name{plot.PCResults}
\alias{plot.PCResults}
\title{Plot the MCMC iterations for each model coefficient.}
\usage{
\method{plot}{PCResults}(object)
}
\arguments{
\item{object}{An object of class \code{PCResults}.  This is the result of
running a model using \code{runPCModel}.}
}
\value{
A plot of the MCMC chains for each model coefficient.
}
\description{
\code{plot.PCResults} generates the MCMC iteration plots for model
coefficients after running a PC model using \code{runPCModel}.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
structs <- chooseStructures(rangs)
PCdat <- createPCData(structs, trimData = FALSE,
                      covariates = c("secondary", "TAM", "secTAM"),
                      covariateTypes = c("binary", "binary", "binary"),
                      covariateLevels = c("sample", "raster", "raster"))
PCmod <- writePCModel(PCdat, multiSampsPerSubj = TRUE, typeOfZero = "censored")
PCresults <- runPCModel(modelObj = PCmod, PCDataObj = PCdat, slideVar='slide',
                        monitorCoefOnly = FALSE,
                        nBurnin = 15000, nIter = 40000, nThin = 25)
plot(PCresults)
}
