% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic functions.R
\name{plot.rangeList}
\alias{plot.rangeList}
\title{Plot the semivariance estimates overlaid with the covariance model}
\usage{
\method{plot}{rangeList}(object)
}
\arguments{
\item{object}{An object of class "rangeList".}
}
\value{
A plot or plots or the semivariance estimates overlaid with the
fitted covariance model.  If there was no spatial variable provided in the
\code{estRange} function then only a single plot is produced.  This is
equivalent to the plots produced by \code{geoR}.  However, if a spatial
variable was provided, then \code{plot.rangeList} produces a plot for
every level of that variable.
}
\description{
\code{plot.rangeList} uses the S3 plot methods in the \code{geoR} package
to plot the semivariance estimates and then overlay the fitted covariance
model.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
plot(rangs)

}
\references{
Cressie, N and Hawkins, DM. 1980. Robust estimation of the
variogram: I. \emph{Journal of the International Association for
Mathematical Geology}, 12(2):115-125.
}
