% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIDs.R
\name{plotIDs}
\alias{plotIDs}
\title{Plot the assigned ID numbers}
\usage{
plotIDs(data, xCoord = "x", yCoord = "y", IDVar = "assignID",
  textSize = 4, TMA = FALSE)
}
\arguments{
\item{data}{The dataset containing x- and y-coordinates and a sample
identifier.}

\item{xCoord}{A character string specifying the name of the x-coordinate
variable.}

\item{yCoord}{A character string specifying the name of the y-coordinate
variable.}

\item{IDVar}{A character string speciffying the sample ID variable.}

\item{textSize}{The size of the sample IDs on the plot.}

\item{TMA}{A TRUE/FALSE variable specifying whether or not the data came
from a tissue microarray (TMA).  This is only used to determine how to
color the samples.  If \code{TMA=TRUE}, then the samples are colored using
a continuous color scale.  If \code{TMA=FALSE}, then the samples are
colored using a discrete color scale.}
}
\value{
A plot of the tissue samples colored according to the sample ID and
overlaid with the ID.
}
\description{
\code{plotIDs} plots the samples and overlays the sample IDs.
}
\examples{
data("TAMdata")
TAMdata <- assignIDs(TAMdata)
plotIDs(TAMdata)
}
