% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRangeObj_final.R
\name{plotRangeObj}
\alias{plotRangeObj}
\title{Plot a range object.}
\usage{
plotRangeObj(rangeObj)
}
\arguments{
\item{rangeObj}{An object of class \code{rangeList}.  This object is a list
that contains the estimated range parameter(s).  This is obtained by a call
to the \code{\link{estRange}} function.}
}
\value{
A set of plots displaying the semivariance estimates, fitted
covariance model, and number of pairs of observations used to estimate
semivariance.  If no spatial variable was provided to the
\code{\link{estRange}} function, then there will be two plots.  The first,
on top, will show the semivariance estimates and the fitted covariance
model.  The second, on the bottom, will show the number of data pairs used
to the estimate the semivariance at the corresponding distance.  If a
spatial variable is supplied to the \code{\link{estRange}} function, then
there will be two plots for every level of the spatial variable.
}
\description{
\code{plotIDs} the semivariance estimates, covariance model, and number of
pairs of observations used to estimate semivariance.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
plotRangeObj(rangs)

}
