% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rScale_final.R
\name{rScale}
\alias{rScale}
\title{Rescale the data}
\usage{
rScale(data, subjectVar = NULL, sampleVar, xCoord, yCoord)
}
\arguments{
\item{data}{The dataset to be rescaled.  The dataset must be a data frame.}

\item{subjectVar}{A character string specifying the subject variable, if it
exists.  This should be specified for paired data only, meaning data in which
there is more than one sample for at least one subject.}

\item{sampleVar}{A character string specifying the sample variable.  This
should be a variable with unique values for each sample.}

\item{xCoord}{A character string specifying the name of the x-coordinate
variable.}

\item{yCoord}{A character string specifying the name of the y-coordinate
variable.}
}
\value{
A list including the rescaled dataset and the names of the subject
and sample variables.  The subject and sample variables are carried forward
to subsequent functions so that they only have to be input once.

\describe{
  \item{\code{data}}{The dataset, appended with new x- and y-coordinates
  to be used in further calculations.}
  \item{\code{subjectVar}}{A character string denoting the subject
  variable.}
  \item{\code{sampleVar}}{A character string denoting the sample variable.}
}
}
\description{
\code{rScale} rescales the coordinates of the data so that the distance
between observations is 1.  It also centers the set of coordinates for each
sample at the origin.
}
\examples{
data("TAMdata")
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
}
