% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic functions.R
\name{summary.PCResults}
\alias{summary.PCResults}
\title{Summarize PCResults object}
\usage{
\method{summary}{PCResults}(object)
}
\arguments{
\item{object}{An object of class "PCResults".  This is the result of a
call to the \code{runPCModel} function.}
}
\value{
The \code{summary.PCResults} function prints the matrix of
summary measures labeled "results" in the object.  This matrix contains
summary information for the model coefficients.
}
\description{
\code{summary.PCResults} provides basic results for an object of class
"PCResults".
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
structs <- chooseStructures(rangs)
PCdat <- createPCData(structs, trimData = FALSE,
                      covariates = c("secondary", "TAM", "secTAM"),
                      covariateTypes = c("binary", "binary", "binary"),
                      covariateLevels = c("sample", "raster", "raster"))
PCmod <- writePCModel(PCdat, multiSampsPerSubj = TRUE, typeOfZero = "censored")
PCresults <- runPCModel(modelObj = PCmod, PCDataObj = PCdat, slideVar='slide',
                        monitorCoefOnly = FALSE,
                        nBurnin = 15000, nIter = 40000, nThin = 25)
summary(PCresults)
}
