% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic functions.R
\name{summary.rangeList}
\alias{summary.rangeList}
\title{Summarize rangeList object}
\usage{
\method{summary}{rangeList}(object)
}
\arguments{
\item{object}{An object of class \code{rangeList}.  This is the result of a
call to the \code{estRange} function.}
}
\value{
The \code{summary.rangeList} prints the estimated range for
every level of the spatial variable provided to the \code{estRange}
function.  If no spatial varaible was provided, then only one estimated
range is printed.
}
\description{
\code{summary.rangeList} provides basic results for an object of class
\code{rangeList}.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
summary(rangs)
}
