% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic functions.R
\name{summary.structureList}
\alias{summary.structureList}
\title{Summarize structureList object}
\usage{
\method{summary}{structureList}(object)
}
\arguments{
\item{object}{An object of class \code{structureList}.  This is the result
of a call to the \code{chooseStructures} function.}
}
\value{
For every level a spatial variable (provided in the \code{estRange}
function), the \code{summary.rangeList} prints the chosen support structure,
the total number of support points across all samples after removing
support points, and the number of samples.  If no spatial variable was
provided, only one line of information is printed.
}
\description{
\code{summary.structureList} provides basic results for an object of class
\code{structureList}.
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
structs <- chooseStructures(rangs)
summary(structs)
}
