% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAMdata.R
\docType{data}
\name{TAMdata}
\alias{TAMdata}
\title{N-glycan abundance measures in primary mammary and secondary lung tumor
tissues.}
\format{A dataframe with 18,154 rows and 59 columns.
\describe{
  \item{Spot.index}{A numeric identifier for each spot or raster.}
  \item{x}{The x-coordinate of each spot.}
  \item{y}{The y-coordinate of each spot.}
  \item{subject}{A numeric subject identifier.}
  \item{ROI}{A numeric identifier for the region of interest (ROI)}
  \item{slide}{A numeric identifier indicating samples that were run on the
  same slide.}
  \item{secondary}{A sample-level binary variable that takes on a value of
  0 if the sample was resected from primary mammary tissue and a value of
  1 if the sample was resected from secondary lung tissue.}
  \item{TAM}{A spot-level binary variable indicating whether the spot
  stained postive (TAM=1) or negative (TAM=0) for tumor-associated
  macrophages (TAMs).}
  \item{meanTAM}{A spot-level numeric variable on a 0-1 scale that
  indicates the proportion of pixels (from a corresponding scanned image)
  within the bounds of the corresponding spot that stained positive for
  TAMs.}
  \item{X771.auc}{The relative abundance, as measured by AUC, of a fragment
  with an m/z of approximately 771.}
  \item{X__.auc}{The relative abundance, as measured by AUC, of a
  fragment with an m/z given by __.}
}}
\usage{
TAMdata
}
\description{
A dataset containing area under the curve (AUC) abundance measures for
50 N-glycans.
}
\keyword{datasets}
