% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePCmodel_final.R
\name{writePCModel}
\alias{writePCModel}
\title{Write the model}
\usage{
writePCModel(PCDataObj, typeOfZero = "censored", covariates = NULL,
  covariatesForBinary = NULL, coefPrior = "sdunif",
  multiSampsPerSubj = NULL, errorVarianceLevel = "sample",
  latentVarianceLevel = "sample", addAssignNode = NULL,
  assignNodeLevel = NULL)
}
\arguments{
\item{PCDataObj}{An object of class \code{PCDataList}.  This object is the
result of using the \code{\link{createPCData}} function.}

\item{typeOfZero}{A character string specifying the type of zero that
will be assumed.  The two options are "censored" (default) and "true".
If there are no zeros for the outcome of interest then this argument
is superfluous.}

\item{covariates}{A concatenated character string specifying the covariates
that will be included in the model.  When there are no zeros in the data,
or when zeros are assumed to be censored, there is only a single set of
covariates.  When the zeros are assumed to be true zeros, there are two
model parts, a binary and a marginalized.  The covariates specified in this
argument will be used in the marginalized part of the model.}

\item{covariatesForBinary}{A concatenated character string specifying
covariates that will be included in the binary part of the marginalized
two part model.  This argument will only be used if there are zeros in
the data and \code{typeOfZero='true'}.  If \code{typeOfZero='true'}, and
this argument is left empty, then the covariates for the binary part of
the model will be assumed to be the same as those for the marginalized part
of the model.}

\item{coefPrior}{A character string specifying the prior that should be
assumed for model coefficients.  The options are "sdunif" (default),
"dnorm", and "Cauchy".  See details for more information.}

\item{multiSampsPerSubj}{A TRUE/FALSE variable indicating whether or not
there are mutliple samples per subject.}

\item{errorVarianceLevel}{A character string specifying the level at which
the error variance should be estimated.  The options are "sample" (default),
"subject", and "overall".  As an example, if
\code{errorVarianceLevel="sample"} then a separate error variance is
estimated for each sample.}

\item{latentVarianceLevel}{A character string specifying the level at which
the latent variance should be estimated.  The options are "sample" (default),
"subject", and "overall".}
}
\value{
A list containing the model and other information
supplied to the \code{\link{createPCData}} function.



\describe{
  \item{\code{model}}{A model object created using the
  \code{link[nimble]{nimbleCode}} function.}
  \item{\code{covariates}}{A concatenated character string of the
  covariate names.}
  \item{\code{covariatesForBinary}}{A concatenated character string of the
  covariate names for the binary part of the marginalized two-part model.}
  \item{\code{coefPrior}}{A character string specifying the assumed prior
  for the model coefficients.}
  \item{\code{multiSampsPerSubj}}{A TRUE/FALSE variable indicating whether
  there are multiple samples per subject.}
  \item{\code{errorVarianceLevel}}{A character string specifying the level
  (overall, subject, or sample) at which the error variance will be
  estimated.}
  \item{\code{latentVarianceLevel}}{A character string specifying the level
  (overall, subject, or sample) at which the latent variance will be
  estimated.}
  \item{\code{typeOfZero}}{A character string specifying the assumption
  made about the type of zeros in the data.}
}
}
\description{
The \code{writePCModel} function writes a NIMBLE model that incorporates the
process convolution approach.
}
\details{
For the model coefficients, three possible priors are allowed.

\strong{sdunif}
For all model coefficients, including the intercept(s), the
prior is given by \deqn{\beta \sim \mbox{N}(0, \sigma^{2}_{\beta});
\sigma_{\beta} \sim \mbox{U}(0,10).}

\strong{dnorm}
For all model coefficients, including the intercept(s), the
prior is given by \deqn{\beta \sim \mbox{N}(0, 0.00001).}

\strong{Cauchy}
The Cauchy priors are specified according to the recommendations of Gelman
et al. (input citation).  While Gelman used the Cauchy priors in the
context of logistic regression, imaging mass spectrometry data is often
lognormally distributed, and on the log scale parameter spaces may need
to be constrained in a similar way.  For intercepts, the prior is
\deqn{\beta_{0} \sim \mbox{Cauchy}(0, \mbox{scale}=10).}
For other model coefficients the prior is
\deqn{\beta \sim \mbox{Cauchy}(0, \mbox{scale}=2.5).}
}
\examples{
data("TAMdata")
# The dataset is trimmed only for the speed of the example
TAMdata <- TAMdata[TAMdata$subject < 3, ]
TAMdata <- rScale(TAMdata, subjectVar = 'subject', sampleVar = 'ROI',
                  xCoord = 'x', yCoord = 'y')
rangs <- estRange(TAMdata, outcome = 'X1282.auc', spatialVar = 'TAM',
                  semivEst = 'modulus', logTransform = TRUE)
structs <- chooseStructures(rangs)
PCdat <- createPCData(structs, trimData = FALSE,
                      covariates = c("secondary", "TAM", "secTAM"),
                      covariateTypes = c("binary", "binary", "binary"),
                      covariateLevels = c("sample", "raster", "raster"))
PCmod <- writePCModel(PCdat, multiSampsPerSubj = TRUE, typeOfZero = "censored")
}
\references{
de Valpine, P., D. Turek, C.J. Paciorek, C. Anderson-Bergman,
D. Temple Lang, and R. Bodik. 2017. Programming with models: writing
statistical algorithms for general model structures with NIMBLE.
\emph{Journal of Computational and Graphical Statistics}. 26: 403-413.
}
